Split out an interface for VisualizationConfigurationFactory
