+++ src/tests/unit/tests/background/tab-context-factory.test.ts
@@ -13,6 +13,7 @@ import { TabContext } from 'background/tab-context';
 import { TabContextFactory } from 'background/tab-context-factory';
 import { TargetTabController } from 'background/target-tab-controller';
 import { TelemetryEventHandler } from 'background/telemetry/telemetry-event-handler';
+import { WebVisualizationConfigurationFactory } from 'common/configs/web-visualization-configuration-factory';
 import { Logger } from 'common/logging/logger';
 import { NotificationCreator } from 'common/notification-creator';
 import { WindowUtils } from 'common/window-utils';
@@ -21,7 +22,6 @@ import { UnifiedScanResultStore } from '../../../../background/stores/unified-sc
 import { UsageLogger } from '../../../../background/usage-logger';
 import { BrowserAdapter } from '../../../../common/browser-adapters/browser-adapter';
 import { VisualizationConfiguration } from '../../../../common/configs/visualization-configuration';
-import { VisualizationConfigurationFactory } from '../../../../common/configs/visualization-configuration-factory';
 import { getStoreStateMessage } from '../../../../common/messages';
 import { PromiseFactory } from '../../../../common/promises/promise-factory';
 import { StoreNames } from '../../../../common/stores/store-names';
@@ -29,7 +29,7 @@ import { StoreUpdateMessage } from '../../../../common/types/store-update-messag
 import { VisualizationType } from '../../../../common/types/visualization-type';
 
 function getConfigs(visualizationType: VisualizationType): VisualizationConfiguration {
-    return new VisualizationConfigurationFactory().getConfiguration(visualizationType);
+    return new WebVisualizationConfigurationFactory().getConfiguration(visualizationType);
 }
 
 describe('TabContextFactoryTest', () => {
